rag-demo/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
faiss-cpu
langchain-openai
langchain-core
python-dotenv